+++ airbyte-integrations/connectors/destination-s3-data-lake/src/main/kotlin/io/airbyte/cdk/load/toolkits/iceberg/parquet/io/Operation.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.destination.s3_data_lake.io
+package io.airbyte.cdk.load.toolkits.iceberg.parquet.io
 
 /** Delta operations for data. */
 enum class Operation {
