+++ airbyte-integrations/connectors/destination-s3-data-lake/src/test/kotlin/io/airbyte/integrations/destination/s3_data_lake/io/BaseDeltaTaskWriterTest.kt
@@ -4,7 +4,9 @@
 
 package io.airbyte.integrations.destination.s3_data_lake.io
 
-import io.airbyte.integrations.destination.s3_data_lake.io.BaseDeltaTaskWriter.RowDataDeltaWriter
+import io.airbyte.cdk.load.toolkits.iceberg.parquet.io.BaseDeltaTaskWriter
+import io.airbyte.cdk.load.toolkits.iceberg.parquet.io.Operation
+import io.airbyte.cdk.load.toolkits.iceberg.parquet.io.RecordWrapper
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
@@ -43,7 +45,7 @@ internal class BaseDeltaTaskWriterTest {
             )
         val primaryKeyIds = setOf(1)
         val schema = Schema(columns, primaryKeyIds)
-        val deltaWriter: RowDataDeltaWriter = mockk {
+        val deltaWriter: BaseDeltaTaskWriter.RowDataDeltaWriter = mockk {
             every { deleteKey(any()) } returns Unit
             every { write(any()) } returns Unit
         }
@@ -93,7 +95,7 @@ internal class BaseDeltaTaskWriterTest {
                 Types.NestedField.required(3, "timestamp", Types.TimestampType.withZone()),
             )
         val schema = Schema(columns)
-        val deltaWriter: RowDataDeltaWriter = mockk {
+        val deltaWriter: BaseDeltaTaskWriter.RowDataDeltaWriter = mockk {
             every { deleteKey(any()) } returns Unit
             every { write(any()) } returns Unit
         }
